Add *first page* concept
The reader view now has a page menu, and a *First Page* checkbox entry
that allows setting the *first page* of the book to the document's
current page. One can then jump to a *book page* by specifying `"n`,
where `n` is book page number, in the *Go to page* input field.