Add support for the Aura H₂O Edition 2
This new device is pressure sensitive, but the pressure value can't be
used to detect a finger release (unlike the other devices where the
pressure is a constant (> 0) on contact/motion and 0 on release).

We therefore keep track of the set of active IDs within a packet and
release the fingers that aren't in the set on each packet sync. event.

The natural light on this device is a combination of white (ledb) and
orange (leda).