Fix the framebuffer rotation on two devices
It would seem that I was misguided by KSM's Plato launch script: the
value that we must write at startup is 3, even for the devices that
exhibit a discrepancy between the written and the read values.

Also, the FBIOPUT_VSCREENINFO ioctl call might fail on the Aura H₂O.
Hence we must check that the read value changes after the call.

The mirroring pattern scheme was generalized to handle the Aura H₂O
Edition 2 Version 2: on this device, an increase of the written rotation
value results in a counter-clockwise rotation of the framebuffer.