Add configuration options for south and west strips
Add "south-strip-action" and "west-strip-action" configuration options
(with possible values "toggle-bars", "next-page" and "previous-page",
"next-page" respectively).
Their default actions retain their previous behavior.

This allows a reader to configure Plato such that a tap along any of the
west, south, or east strips advances to the next page.  This allows that
common action to be performed easily irrespective of how the reader is
holding their device.

Note: if the west strip is configured to go to the next page then reader
can still navigate to the previous page with a easterly swipe.