Turn off LED even when not in standalone mode
Disable the charging indicator LED even if Plato isn't being run in
standalone mode.

This ensures that even if the device was charging when Plato was
started, the LED will still be turned off.  This is important because
Plato will not turn the LED off itself neither when charging is
complete nor when the USB cable is unplugged.

Fixes: https://github.com/baskerville/plato/issues/175