Add HiDPI support for the emulator
Enable high DPI support in the SDL2 window, and scale the window to show
the emulated framebuffer in a pixel-perfect way. That means a device
like the Touch A/B will be displayed at 200% size, while many others
like the Clara BW will be displayed at 100% size.

This only has an effect with SDL_VIDEODRIVER=wayland.